@@ -1,14 +1,16 @@
 import { CommonModule, DatePipe, NgClass, NgFor, NgIf } from '@angular/common';
-import { Component, AfterViewInit, NgModule, ElementRef, OnInit, HostListener } from '@angular/core';
+import { Component, AfterViewInit, NgModule, ElementRef, OnInit, HostListener, viewChild, ViewChild, ViewChildren } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { Route, Router, RouterModule } from '@angular/router';
-import {provideNativeDateAdapter} from '@angular/material/core';
-import {MatDatepickerModule} from '@angular/material/datepicker';
-import {MatInputModule} from '@angular/material/input';
-import {MatSelectModule} from '@angular/material/select';
-import {MatFormFieldModule} from '@angular/material/form-field';
-import {MatSliderModule} from '@angular/material/slider';
+import { Router, RouterModule } from '@angular/router';
+import { provideNativeDateAdapter } from '@angular/material/core';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatInputModule } from '@angular/material/input';
+import { MatSelectModule } from '@angular/material/select';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatSliderModule } from '@angular/material/slider';
 import { DateAdapter } from '@angular/material/core';
+import { TranslateModule, TranslatePipe, TranslateService } from '@ngx-translate/core';
+import { MatSelect } from '@angular/material/select';
 
 
 
@@ -23,13 +25,13 @@ interface Source {
   selector: 'app-main',
   standalone: true,
   providers: [provideNativeDateAdapter()],
-  imports: [MatSliderModule,MatFormFieldModule,MatSelectModule, MatInputModule, FormsModule,NgClass, NgFor, NgIf, FormsModule,CommonModule,RouterModule, FormsModule,MatFormFieldModule, MatInputModule, MatDatepickerModule],
+  imports: [MatSliderModule, MatFormFieldModule, MatSelectModule, MatInputModule, FormsModule, NgClass, NgFor, NgIf, FormsModule, CommonModule, RouterModule, FormsModule, MatFormFieldModule, MatInputModule, MatDatepickerModule, TranslateModule,],
   templateUrl: './main.component.html',
   styleUrl: './main.component.css'
 })
 
 
-export class MainComponent implements OnInit{
+export class MainComponent implements OnInit {
 
 
   //ინფორმაცია სორსების ასარჩევად
@@ -38,29 +40,69 @@ export class MainComponent implements OnInit{
   }
 
 
-  //ინფორმაცია სორსების ასარჩევად
-  sources: Source[] = [
-    {value: 'თბილისი', viewValue: 'თბილისი'},
-    {value: 'ბათუმი', viewValue: 'ბათუმი'},
-    {value: 'ფოთი', viewValue: 'ფოთი'},
-  ];
 
 
 
-  
   source: string = '';
   destination: string = '';
   date: string = '';
   travelers: number = 1;
   filteredDestinations: Source[] = [];
+  language: string = ''
 
-  constructor(private router: Router, private dateAdapter: DateAdapter<Date>) {
+  //ინფორმაცია სორსების ასარჩევად
+
+  sources: Source[] = [
+    { value: 'თბილისი', viewValue: this.language == 'geo' ? 'თბილისი' : 'Tbilisi' },
+    { value: 'ბათუმი', viewValue: 'ბათუმი' },
+    { value: 'ფოთი', viewValue: 'ფოთი' },
+  ];
+
+  updateSourcesViewValue() {
+    this.language = JSON.parse(localStorage.getItem('selectedLanguage') || '');
+    switch (this.language) {
+      case 'geo':
+        this.sources[0].viewValue = 'თბილისი';
+        this.sources[1].viewValue = 'ბათუმი';
+        this.sources[2].viewValue = 'ფოთი';
+        break;
+      default:
+        // If language is not 'geo', set default values
+        this.sources[0].viewValue = 'Tbilisi';
+        this.sources[1].viewValue = 'Batumi';
+        this.sources[2].viewValue = 'Poti';
+
+        break;
+    }
+  }
+
+
+
+  onSelectMenuOpen(isOpened: boolean) {
+    if (isOpened) {
+        // Always fetch and update the language and view values whenever the menu is opened
+        this.language = JSON.parse(localStorage.getItem('selectedLanguage') || '{}');
+        console.log("header component lang", this.language);
+        this.updateSourcesViewValue();
+    }
+}
+
+
+  constructor(private router: Router, private dateAdapter: DateAdapter<Date>, private translate: TranslateService) {
+    translate.setDefaultLang('en');
+
+
+
     // Setting the locale to be used by date adapter
     this.dateAdapter.setLocale('en-GB'); // Set your desired locale
+
   }
 
   ngOnInit() {
     this.filteredDestinations = [...this.sources];
+
+
+
   }
 
 
@@ -140,7 +182,4 @@ export class MainComponent implements OnInit{
 
 
 
- 
 }
-
-
