@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { RouterModule, } from '@angular/router';
 
 import { FormBuilder, FormGroup, FormArray, Validators, ReactiveFormsModule } from '@angular/forms';
+import { TranslateModule, TranslatePipe, TranslateService } from '@ngx-translate/core';
 
 
 
@@ -9,14 +10,19 @@ import { FormBuilder, FormGroup, FormArray, Validators, ReactiveFormsModule } fr
 @Component({
   selector: 'app-pay',
   standalone: true,
-  imports: [RouterModule],
+  imports: [RouterModule,TranslateModule],
   templateUrl: './pay.component.html',
   styleUrl: './pay.component.css'
 })
 export class PayComponent  {
   ticketId: string | null=null;
+  selectedLanguage: string = 'en'; 
 
-  constructor() { }
+  constructor(private translateService: TranslateService) {
+    this.translateService.setDefaultLang(this.selectedLanguage);
+
+
+  }
 
 onSubmit(){
   // this.ticketId = sessionStorage.getItem('ticketId') || null;
